+++ esibd/devices/lakeshore335/lakeshore.py
@@ -89,7 +89,7 @@ class LS335(Device):
     def convertDataDisplay(self, data: np.ndarray) -> np.ndarray:
         return data - 273.15 if self.unitAction.state else data
 
-    def getUnit(self) -> str:
+    def getDisplayUnit(self) -> str:
         return '°C' if self.unitAction.state else self.unit
 
     def updateTheme(self) -> None: